requirements.txt
aioredis==1.1.0
aliyun-python-sdk-core-v3==2.9.1
amqp==2.3.2
ansible==2.6.4
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
channels==2.1.1
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
daphne==2.1.1
Django==2.0.5
django-auth-ldap==1.7.0
django-celery-beat==1.1.1
django-cors-headers==2.4.0
django-filter==1.1.0
django-mysql==2.4.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dnspython==1.15.0
docopt==0.6.2
ecdsa==0.13
future==0.17.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
msgpack==0.5.6
mysqlclient==1.3.13
paramiko==2.4.2
Pillow==5.2.0
prompt-toolkit==2.0.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodome==3.6.6
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
PyNaCl==1.2.1
pyotp==2.2.6
python-ldap==3.1.0
pytz==2018.5
pyvim==2.0.22
pyvmomi==6.7.0.2018.9
PyYAML==3.13
qingcloud-sdk==1.2
qiniu==7.2.2
qrcode==6.0
redis==2.10.6
requests==2.19.1
six==1.11.0
sshpubkeys==3.1.0
Twisted==18.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
zope.interface==4.5.0